chore: cleanup resolver export
